FEAT: 프로젝트 배포

See merge request s08-final/S08P31B308!248

Former-commit-id: 0c9aeba48a082e28a2a24a9e978cb6989603392e

+++ another_front/lib/screens/running/widgets/distancebar_custom.dart
@@ -1,10 +1,10 @@
-import 'package:another/constant/color.dart';
+import 'package:another/constant/const/color.dart';
 import 'package:flutter/material.dart';
 
 class CustomSliderThumbCircle extends SliderComponentShape {
   final double thumbRadius;
-  final int min;
-  final int max;
+  final double min;
+  final double max;
 
   const CustomSliderThumbCircle({
     required this.thumbRadius,
